requirements.txt
streamlit
openai
arxiv
pymupdf
python-dotenv
Pillow
requests
qrcode